Split relational storage catalog
