+++ crates/truthdb-core/src/storage/relational/catalog/mod.rs
@@ -1,0 +1,4 @@
+mod databases;
+mod indexes;
+mod routines;
+mod schema_objects;
